Update 9. PROGRAM TO READ TIME IN HR,MIN,SEC AND CONVERT IT INTO TOTAL SECOND.
made a lot of changes, the user can now choose any 1 format for time out of two. Before this the user is not aware about the format, before this if the user would give the input of the 12-hour format there would be a logic error and the user can't specify A.M or P.M. In this the user can choose between the 12-hour format and 24-hour format.